# [DEVOPS-391] Remove gulp; Babel 6 --> 7