--- a/.gitignore
+++ b/.gitignore
@@ -27,6 +27,7 @@ dist
 lib-cov
 
 # Coverage directory used by tools like istanbul
+.nyc_output/
 coverage
 
 # Grunt intermediate storage (http://gruntjs.com/creating-plugins#storing-task-files)