--- a/.sequelizerc
+++ b/.sequelizerc
@@ -1,8 +1,8 @@
 const path = require('path');
 
 module.exports = {
-  'config': path.resolve('config', 'database.js'),
-  'models-path': path.resolve('server', 'models'),
-  'seeders-path': path.resolve('server', 'seeders'),
-  'migrations-path': path.resolve('server', 'migrations')
+  'config': path.resolve('src', 'config', 'database.js'),
+  'models-path': path.resolve('src', 'db', 'models'),
+  'seeders-path': path.resolve('src', 'db', 'seeders'),
+  'migrations-path': path.resolve('src', 'db', 'migrations')
 }