--- a/package.json
+++ b/package.json
@@ -11,27 +11,27 @@
     "yarn": ">=0.20.3"
   },
   "scripts": {
-    "start": "gulp serve",
-    "serve": "node dist/index.js",
-    "migrate": "./node_modules/.bin/sequelize db:migrate",
-    "migrate:undo": "./node_modules/.bin/sequelize db:migrate:undo:all",
+    "build": "babel create_db.js delete_db.js src --out-dir dist",
     "create_db": "cross-env NODE_ENV=test node dist/create_db.js",
     "delete_db": "cross-env NODE_ENV=test node dist/delete_db.js",
-    "start:debug": "cross-env DEBUG=amida-auth-microservice:* yarn start",
-    "build": "gulp",
+    "clean": "rimraf coverage dist",
+    "jenkins": "NODE_ENV=test yarn delete_db && NODE_ENV=test yarn create_db && NODE_ENV=test yarn migrate && NODE_ENV=test yarn test",
     "lint": "esw -c .eslintrc *.js server config . --color --fix",
     "lint:watch": "yarn lint -- --watch",
-    "beautify": "esw -c .eslintrc *.js server config --color --fix",
+    "migrate": "sequelize db:migrate",
+    "migrate:undo": "sequelize db:migrate:undo:all",
     "precommit": "yarn lint && yarn test",
-    "test": "gulp && yarn create_db && cross-env NODE_ENV=test yarn migrate && cross-env NODE_ENV=test ./node_modules/.bin/mocha --ui bdd --reporter spec --colors --compilers js:babel-core/register server/tests --recursive -t 10000 && yarn delete_db",
-    "jenkins": "gulp && yarn delete_db && yarn create_db && cross-env NODE_ENV=test yarn migrate && cross-env NODE_ENV=test ./node_modules/.bin/mocha --ui bdd --reporter spec --colors --compilers js:babel-core/register server/tests --recursive -t 10000",
-    "test:watch": "yarn test -- --watch",
-    "test:coverage": "cross-env NODE_ENV=test ./node_modules/.bin/istanbul cover _mocha -- --ui bdd --reporter spec --colors --compilers js:babel-core/register server/tests --recursive",
-    "test:check-coverage": "yarn test:coverage && istanbul check-coverage",
-    "report-coverage": "coveralls < ./coverage/lcov.info",
-    "snyk-protect": "snyk protect",
     "prepublish": "npm run snyk-protect",
-    "seed": "gulp && cross-env DEBUG=amida-auth-microservice:* node dist/config/seed.js && gulp clean"
+    "report-coverage": "coveralls < ./coverage/lcov.info",
+    "seed": "yarn build && cross-env DEBUG=amida-auth-microservice:* node dist/config/seed.js",
+    "serve": "node dist/index.js",
+    "snyk-protect": "snyk protect",
+    "start": "nodemon --exec babel-node -- src/index.js",
+    "start:debug": "cross-env DEBUG=amida-auth-microservice:* yarn start",
+    "test": "NODE_ENV=test mocha --ui bdd --reporter spec --colors --compilers js:@babel/register tests --recursive -t 10000",
+    "test:check-coverage": "yarn test:coverage && istanbul check-coverage",
+    "test:coverage": "cross-env NODE_ENV=test istanbul cover _mocha -- --ui bdd --reporter spec --colors --compilers js:babel-core/register src/tests --recursive",
+    "test:watch": "yarn test -- --watch"
   },
   "dependencies": {
     "bluebird": "3.4.6",
@@ -68,8 +68,7 @@
     "pg-hstore": "2.3.2",
     "rand-token": "^0.4.0",
     "run-sequence": "^1.1.5",
-    "sequelize": "^4.28.0",
-    "sequelize-cli": "^4.1.1",
+    "sequelize": "^4.40.0",
     "snyk": "^1.73.0",
     "supertest": "^3.0.0",
     "swagger-stats": "^0.95.6",
@@ -77,11 +76,11 @@
     "winston": "^3.0.0"
   },
   "devDependencies": {
-    "babel-cli": "6.18.0",
-    "babel-core": "6.18.0",
-    "babel-plugin-add-module-exports": "0.2.1",
-    "babel-preset-es2015": "6.16.0",
-    "babel-preset-stage-2": "6.18.0",
+    "@babel/cli": "^7.1.5",
+    "@babel/core": "^7.2.0",
+    "@babel/node": "^7.0.0",
+    "@babel/preset-env": "^7.1.6",
+    "@babel/register": "^7.0.0",
     "chai-datetime": "1.5.0",
     "codacy-coverage": "2.0.2",
     "coveralls": "^2.11.6",
@@ -97,16 +96,10 @@
     "gulp-util": "^3.0.7",
     "husky": "^0.13.1",
     "istanbul": "1.1.0-alpha.1",
-    "mocha": "3.2.0"
-  },
-  "babel": {
-    "presets": [
-      "es2015",
-      "stage-2"
-    ],
-    "plugins": [
-      "add-module-exports"
-    ]
+    "mocha": "3.2.0",
+    "nodemon": "^1.18.0",
+    "rimraf": "^2.6.0",
+    "sequelize-cli": "^5.4.0"
   },
   "snyk": true
 }