--- a/package.json
+++ b/package.json
@@ -11,9 +11,9 @@
     "yarn": ">=0.20.3"
   },
   "scripts": {
-    "build": "babel create_db.js delete_db.js src --out-dir dist",
-    "create_db": "cross-env NODE_ENV=test node dist/create_db.js",
-    "delete_db": "cross-env NODE_ENV=test node dist/delete_db.js",
+    "build": "babel src --out-dir dist",
+    "create_db": "sequelize db:create",
+    "delete_db": "sequelize db:drop",
     "clean": "rimraf coverage dist",
     "jenkins": "NODE_ENV=test yarn delete_db && NODE_ENV=test yarn create_db && NODE_ENV=test yarn migrate && NODE_ENV=test yarn test",
     "lint": "esw -c .eslintrc *.js server config . --color --fix",