--- a/package.json
+++ b/package.json
@@ -81,6 +81,7 @@
     "@babel/node": "^7.0.0",
     "@babel/preset-env": "^7.1.6",
     "@babel/register": "^7.0.0",
+    "babel-plugin-source-map-support": "^2.0.0",
     "chai-datetime": "1.5.0",
     "codacy-coverage": "2.0.2",
     "coveralls": "^2.11.6",
@@ -99,7 +100,8 @@
     "nodemon": "^1.18.0",
     "nyc": "^13.0.0",
     "rimraf": "^2.6.0",
-    "sequelize-cli": "^5.4.0"
+    "sequelize-cli": "^5.4.0",
+    "source-map-support": "0.5.9"
   },
   "snyk": true
 }