--- a/package.json
+++ b/package.json
@@ -36,13 +36,10 @@
   "dependencies": {
     "bluebird": "3.4.6",
     "body-parser": "1.18.2",
-    "chai": "^3.4.1",
     "compression": "1.7.1",
     "cookie-parser": "1.4.3",
     "cors": "2.8.1",
-    "del": "^2.2.0",
     "dotenv": "^4.0.0",
-    "eslintrc": "^1.0.6",
     "express": "4.16.0",
     "express-actuator": "^1.1.0",
     "express-jwt": "5.3.0",
@@ -66,8 +63,6 @@
     "pg-hstore": "2.3.2",
     "rand-token": "^0.4.0",
     "sequelize": "^4.40.0",
-    "snyk": "^1.73.0",
-    "supertest": "^3.0.0",
     "swagger-stats": "^0.95.6",
     "uuid": "3.1.0",
     "winston": "^3.0.0"
@@ -79,6 +74,7 @@
     "@babel/preset-env": "^7.1.6",
     "@babel/register": "^7.0.0",
     "babel-plugin-source-map-support": "^2.0.0",
+    "chai": "^3.4.1",
     "chai-datetime": "1.5.0",
     "codacy-coverage": "2.0.2",
     "coveralls": "^2.11.6",
@@ -93,7 +89,9 @@
     "nyc": "^13.0.0",
     "rimraf": "^2.6.0",
     "sequelize-cli": "^5.4.0",
-    "source-map-support": "0.5.9"
+    "snyk": "^1.73.0",
+    "source-map-support": "0.5.9",
+    "supertest": "^3.0.0"
   },
   "snyk": true
 }