--- a/src/db/seeders/20181211041630-create-admin-user.js
+++ b/src/db/seeders/20181211041630-create-admin-user.js
@@ -1,4 +1,4 @@
-'use strict';
+
 
 /**
  * Populate DB with sample data on server start
@@ -6,7 +6,8 @@
  */
 
 module.exports = {
-  up: (queryInterface, Sequelize) => {
+    // eslint-disable-next-line no-unused-vars
+    up: (queryInterface, Sequelize) => {
     /** FIXME: allow three options
      * - reseed everything (nuke)
      * - if there is any User data, seed, but don't overwrite the existing user(s)
@@ -18,17 +19,16 @@ module.exports = {
      * print it to the log, and force the admin to re-auth
      */
 
-    const adminUser = {
-        username: 'admin',
-        email: 'admin@default.com',
-        password: 'admin',
-        scopes: ['admin'],
-    };
+        const adminUser = {
+            username: 'admin',
+            email: 'admin@default.com',
+            password: 'admin',
+            scopes: ['admin'],
+        };
 
-    return queryInterface.bulkInsert('Users', [adminUser]);
-  },
+        return queryInterface.bulkInsert('Users', [adminUser]);
+    },
 
-  down: (queryInterface, Sequelize) => {
-    return queryInterface.bulkDelete('Users', null, {});
-  }
+    // eslint-disable-next-line no-unused-vars
+    down: (queryInterface, Sequelize) => queryInterface.bulkDelete('Users', null, {}),
 };