If this PR fixes a bug, you _must_ add test cases representative of the bug.

#### What's this PR do?
##### from CHANGELOG
- Babel6 --> 7
- gulp3
  * ~5x faster `yarn build` 
  * Identical functionality (`yarn start` for nodemon, `yarn clean`, sourcemaps)
- `create_db.js` and `delete_db.js`
  * Unnecessary because Sequelize includes `sequelize db:create` and `sequelize db:drop`

##### Additional review notes
- We now have proper migrations and seeds in Sequelize
  * Removed `create_db.js` and `delete_db.js`. Unnecessary because Sequelize includes `sequelize db:create` and `sequelize db:drop`
  * `yarn seed` now uses Sequelize
- config support full ES6; no need to use require anymore

##### ~5x faster build time

Now
```
$ time yarn build
yarn run v1.12.1

babel src --out-dir dist
Successfully compiled 27 files with Babel.
Done in 0.99s.

real	0m1.173s
```

Before
```
$ time yarn build

yarn run v1.9.4
gulp
[02:17:31] Requiring external module babel-register
[02:17:31] Using gulpfile ~/git/amida-tech/amida-auth-microservice/gulpfile.babel.js
[02:17:31] Starting 'clean'...
[02:17:31] Finished 'clean' after 31 ms
[02:17:31] Starting 'default'...
[02:17:31] Starting 'copy'...
[02:17:31] Starting 'babel'...
[02:17:31] Finished 'default' after 122 ms
[02:17:36] Finished 'copy' after 4.57 s
[02:17:36] Finished 'babel' after 5.06 s
Done in 5.99s.

real	0m6.277s
```

##### Can use any node version
This PR also resolves ORANGE-794
A bug where we were restricted to max node v8

#### Related JIRA tickets:

#### How should this be manually tested?
Execute all of the commands in `package.json` w/ `yarn` using node >v8

#### Any background context you want to provide?

#### Screenshots (if appropriate):