@rmharrison 2018-12-19
In the other services, I'm leaning towards breaking-up into several smaller sequential PRs:
- Sequelize
- Remove gulp
- babel 6-->7