@unprofessional 2019-01-11
- Functionality is good
- POSTing to the API looks good
- Regression testing via Indaba service suite looks good

Do we want to the default user for the test DB process(es) to be `jenkins`? Understanding that these are meant to be tested on a Jenkins instance after a webhook catches a merge or something. What is a local-dev to do?

- Can we add a new script for local testing? i.e. `yarn test:dev` or `yarn test:local` or something?

`docker-compose` isn't working properly for me:
```
ERROR: for postgres  Cannot start service postgres: b'driver failed programming external connectivity on endpoint amida-auth-microservice_postgres_1 (e0dbbf4e1ee3a76097f219d4c3836bda342d6111212f18f547835eff22df5698): Error starting userland proxy: Bind for 0.0.0.0:5432 failed: port is already allocated'
```
 I suspect because of unrelated crap I've been doing on my own involving postgres -- though I'm not exactly sure why docker containers care about stuff running off of localhost.... I blame their poor networking support for OS X. I'll try it again tomorrow after I regain the energy to kill my services etc.