@rmharrison 2019-01-11
Thanks for reviewing @unprofessional.

## Test script
> Can we add a new script for local testing? i.e. yarn test:dev or yarn test:local or something?
==> 
> Realized we can just modify our local .env.test to suit our local-dev needs.

You're content with using `.env.test` for both local and jenkins. I'd considered created a `.env.jenkins`, but that makes 3x `.env.*` files to maintain, and at this point there wouldn't be any substantive differences between `.env.test` and `.env.jenkins`

For what it's worth, I *hate* the default names in `.env.test`
> AUTH_SERVICE_PG_USER='jenkins'
> AUTH_SERVICE_PG_PASSWORD='jenkins'

I'd prefer `postgres:admin`. Easy fix, but requires changing the Jenkins build server; ideally for all the microservice repos to keep it consistent.

## Postgres
> I suspect because of unrelated crap I've been doing on my own involving postgres

Probably. Check with `ps -ef | grep postgres`

You can also run:
```
docker run -p 5432:5432 -d -e POSTGRES_DB=auth-test -e POSTGRES_USER=jenkins -e POSTGRES_PASSWORD=jenkins postgres:9.6
```

Which should fail with the same error should you have another local postgres running on `port 5432`.