@dbrezack 2019-01-24
@rmharrison Changes look great, having an issue with the yarn migrate command. Looking into fixing it now

```
ERROR: syntax error at or near "NOT"

error Command failed with exit code 1.
```