@rmharrison 2019-01-24
@dbrezack You're using an old version of postgres. Please use 9.6+ (not 9.4.11).

Fixed in the `docker-compose.yml` in another PR, https://github.com/amida-tech/amida-auth-microservice/pull/83