@rmharrison 2019-01-25
@dbrezack Please double-check the `docker-compose.yml`. I caught the 9.4 --> 9.6 bug and opened a separate PR (recommend merging https://github.com/amida-tech/amida-auth-microservice/pull/83 first to avoid duplication), but there may be other lurking issues I didn't catch, e.g. Ubuntu (me) vs Mac (everyone else).