@puckybreg 2019-01-29
@rmharrison I know the actual migration files all pre-date this PR, but something we should maybe make a future ticket for is writing the undo portion of all of these migrations. As it stands several of them are just `return true`.