@rmharrison 2019-01-29
@puckybreg : You're correct, 3/5 migrations are not complete, i.e. no `down` for:
- 20180904132710-create-user.js
- 20180904132711-create-refreshToken.js
- 0181114172213-userId-refreshToken.js

PR #80 was intended as a restructuring PR. Not updating the migrations was intentional.

|                                          | up | down |
|------------------------------------------|----|------|
| 20180904132710-create-user.js            | X  |      |
| 20180904132711-create-refreshToken.js    | X  |      |
| 20180904132721-add-uuid-column-users.js  | X  | X    |
| 20180907172213-uuid-index-user.js        | X  | X    |
| 20181114172213-userId-refreshToken.js    | X  |      |