@rmharrison 2019-01-30
@mountHouli : Noticed the `label:hold` on https://github.com/amida-tech/amida-notification-microservice/pull/21

Did you also want to release the auth service for the Orange release without this refactoring PR?

If so, I would create auth-service `release/2.7.0` before merging this PR to `develop`, including:
- bug/postgres-from-9.4-to-9.6
- feature/DEVOPS-390-husky-pre-push

If not, I would merge to `branch:develop` and you can cut the release next week.